% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSSS.R
\name{BSSS}
\alias{BSSS}
\title{Blind Source Separation Over Space}
\usage{
BSSS(x, coord, kernel_type, kernel_parameter, kernel_list = NULL)
}
\arguments{
\item{x}{A numeric matrix of dimension c(n, p), where the p columns correspond to the entries of the random field and the n rows are the observations.}

\item{coord}{A numeric matrix of dimension c(n,2) where each row represents the coordinates of a point in the spatial domain. Only needed if the argument kernel_list is NULL.}

\item{kernel_type}{A string indicating which kernel function to use. Either 'ring', 'ball' or 'gauss'.}

\item{kernel_parameter}{A numeric vector that gives the parameters for the kernel function. At least length of one for 'ball' and 'gauss' or two for 'ring' kernel.}

\item{kernel_list}{List of spatial kernel matrices with dimension c(n,n). Can be computed by the function \code{\link[SpatialBSS]{spatial_kernel_matrix}}.}
}
\value{
BSSS returns a list, including the estimation of maxing matrix, and the estimated latent field.
}
\description{
BSSS estimates the mixing matrix of blind source separation model for multivariate spatial data.
}
\details{
BSSS estimates the mixing matrix by combining the information of all local covariance matrices together and conduct eigenanalysis.
}
