% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_score.R
\name{d_score}
\alias{d_score}
\title{d score}
\usage{
d_score(estimator, true_value)
}
\arguments{
\item{estimator}{A square matrix, usually an estimator of the \code{true_value} matrix.}

\item{true_value}{A square matrix, which the estimator is compared to.}
}
\value{
A numeric value in [0,1].
}
\description{
d score measures the similarity of two square matrix with same dimension. d_score equals 0 if the estimator is a column permutation of true value.
}
