% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_matern_gaussian_rf.R
\name{gen_matern_gaussian_rf}
\alias{gen_matern_gaussian_rf}
\title{Generating Gaussian random fields with Matern covariance function}
\usage{
gen_matern_gaussian_rf(coords, dim, nu, kappa)
}
\arguments{
\item{coords}{coordinate of target randon field to be generated}

\item{dim}{dimension of target randon field to be generated}

\item{nu}{parameter of matern covariance function}

\item{kappa}{parameter of matern covariance function}
}
\value{
A data matrix with number of rows equal to `coords`, and number of columns equal to `dim`.
}
\description{
Generate Gaussian random fields with Matern covariance function
}
